Add support for parsing Request messages
To be able to parse Request messages some state needs to be kept as the
request objects don't indicate their type only their position as per the
last source capabilities.

Add a simple trait for maintaining pdo state that current just
translates between the object position of the request type and the
matching type. This trait is for SourceCapabilities (potentially wrapped
in an Option) which allows implementations to maintain state simply by
saving the last SourceCapabilities message.