PDO: Add utility functions to translate voltage/current/power
The voltage/current/power values in the various pdos require
translations to real numbers; Add utility functions for this and rename
the field to raw_ to avoid accidentally using the wrong function.

For clarity also suffix the function by the returned unit if it's not
the standard unit.

Also for consistency use max instead of maximum everywhere

Signed-off-by: Sjoerd Simons <sjoerd@collabora.com>